do not destructure from `context`
While refactoring, I ran into issues because sometimes we destructure
`parent` which happens to be a "global" value that TypeScript is fine
with if it's undefined.

This will make future commits a bit more readable as well.

Might go back to a destructured pattern, but let's do this for now.